Distill Aggregation object, refactor tests
This solution scauses a serialization error, the text of which is stored
in .serializable_err.txt. We encounter this error because the
generateCrossSectional function is private - something with the
PrivateMethodTester work around triggers the error.

We should fix this by removing the generateCrossSectional function and
making a public constructor for the CrossSectional class. This is part
of a larger effort which will be handled later this quarter (Bug 1298123:
Refactor dataset classes so they do not use case classes).