blocking error in setting root.set
setting a key in root requires the key be in the schema. To add all
child types to the schema, we need a whitelist of types when building
hystogram.definitions.

To fix this, add multiple entries for each histogram definition. Then
build a function that looks in all of these places, similar to our first
solution. will require a whitelist.